implemented batch fetch configured using jdbc.fetch.size

+++ integration-tests/src/test/resources/sql/hsqldb/drop.table.all-types.sql
@@ -1,1 +1,1 @@
-drop table test_all_types
+drop table test_all_types if exists
